requirements.txt
streamlit
sentence-transformers
faiss-cpu
pymupdf
requests
pandas
numpy
python-dotenv